Add files via upload
This is a Tableau Workbook that displays a map of CO2 emissions per country across the years 2000 to 2011. A filter can be used to show on the map only the CO2 emissions per country for a particular year between 2000 and 2011. Four datasets were connected using inner joins to enable the creation of the map.